Initial roughed in project scaffolding and notes.
The basic directory structure is setup, as are the testing
dependencies (phpcs and phpunit) that'll be needed later.

The stated challenge spec was copied over to a txt file in docs/, and
docs/running_tod.md is serving as a project planning board.

The README.md structure is filling in, and will end up being the
official documentation once this is complete.