Built most of a generic HTTP framework.
Developed a set of Request and Response objects, to provide
simple http request and response behavior.

Roughed in a project Router and Controller structure, and provided
a few exceptions to capture typical error behavior.

More work on the __main.php file, to set up and combine the principle
pieces of the framework.

Added a typical nginx config file.